Make translations functions server only
