+++ apps/consent-ui/package.json
@@ -28,6 +28,7 @@
 		"react-hook-form": "^7.47.0",
 		"react-select": "^5.8.0",
 		"sass": "^1.64.0",
+		"server-only": "^0.0.1",
 		"sharp": "^0.32.6",
 		"tailwindcss": "3.3.2",
 		"types": "workspace:^",
