+++ apps/consent-ui/src/i18n/utils/index.ts
@@ -18,3 +18,4 @@
  */
 
 export * from './getTranslation';
+export * from './replaceParams';
